+++ src/egon/data/datasets/sanity_checks.py
@@ -1,26 +1,25 @@
 """
-This module does sanity checks for both the eGon2035 and the eGon100RE scenario seperately where a percentage
-error is given to showcase difference in output and input values. Please note that there are missing input technologies in the supply tables.
- Authors: @ALonso, @dana
+This module does sanity checks for both the eGon2035 and the eGon100RE scenario
+separately where a percentage error is given to showcase difference in output
+and input values. Please note that there are missing input technologies in the
+supply tables.
+Authors: @ALonso, @dana
 """
 from math import isclose
 from pathlib import Path
-from loguru import logger
-import matplotlib.pyplot as plt
-import seaborn as sns
 
 from sqlalchemy import Numeric
 from sqlalchemy.sql import and_, cast, func, or_
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import seaborn as sns
 
-import egon.data
-from egon.data import db
+from egon.data import db, logger
 from egon.data.datasets import Dataset
-from egon.data.datasets.power_plants.pv_rooftop_buildings import (
-    municipality_data,
-    osm_buildings,
-    scenario_data,
+from egon.data.datasets.electricity_demand_timeseries.cts_buildings import (
+    EgonCtsElectricityDemandBuildingShare,
+    EgonCtsHeatDemandBuildingShare,
 )
 from egon.data.datasets.emobility.motorized_individual_travel.db_classes import (
     EgonEvCountMunicipality,
@@ -42,25 +41,35 @@ from egon.data.datasets.etrago_setup import (
     EgonPfHvStore,
     EgonPfHvStoreTimeseries,
 )
+from egon.data.datasets.power_plants.pv_rooftop_buildings import (
+    municipality_data,
+    osm_buildings,
+    scenario_data,
+)
 from egon.data.datasets.scenario_parameters import get_sector_parameters
+import egon.data
 
 
 class SanityChecks(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="SanityChecks",
-            version="0.0.4",
+            version="0.0.5",
             dependencies=dependencies,
             tasks={
-                sanitycheck_eGon2035_electricity,
-                sanitycheck_eGon2035_heat,
                 sanitycheck_pv_rooftop_buildings,
                 sanitycheck_emobility_mit,
+                etrago_eGon2035_electricity,
+                etrago_eGon2035_heat,
+                residential_electricity_annual_sum,
+                residential_electricity_hh_refinement,
+                cts_electricity_demand_share,
+                cts_heat_demand_share,
             },
         )
 
 
-def sanitycheck_eGon2035_electricity():
+def etrago_eGon2035_electricity():
     """Execute basic sanity checks.
 
     Returns print statements as sanity checks for the electricity sector in
@@ -78,9 +87,10 @@ def sanitycheck_eGon2035_electricity():
     scn = "eGon2035"
 
     # Section to check generator capacities
-    print(f"Sanity checks for scenario {scn}")
-    print(
-        "For German electricity generators the following deviations between the inputs and outputs can be observed:"
+    logger.info(f"Sanity checks for scenario {scn}")
+    logger.info(
+        "For German electricity generators the following deviations between "
+        "the inputs and outputs can be observed:"
     )
 
     carriers_electricity = [
@@ -105,7 +115,8 @@ def sanitycheck_eGon2035_electricity():
                         SELECT bus_id FROM grid.egon_etrago_bus
                         WHERE scn_name = 'eGon2035'
                         AND country = 'DE')
-                    AND carrier IN ('biomass', 'industrial_biomass_CHP', 'central_biomass_CHP')
+                    AND carrier IN ('biomass', 'industrial_biomass_CHP',
+                    'central_biomass_CHP')
                     GROUP BY (scn_name);
                 """,
                 warning=False,
@@ -113,7 +124,8 @@ def sanitycheck_eGon2035_electricity():
 
         else:
             sum_output = db.select_dataframe(
-                f"""SELECT scn_name, SUM(p_nom::numeric) as output_capacity_mw
+                f"""SELECT scn_name,
+                 SUM(p_nom::numeric) as output_capacity_mw
                          FROM grid.egon_etrago_generator
                          WHERE scn_name = '{scn}'
                          AND carrier IN ('{carrier}')
@@ -141,25 +153,27 @@ def sanitycheck_eGon2035_electricity():
             sum_output.output_capacity_mw.sum() == 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
-                f"No capacity for carrier '{carrier}' needed to be distributed. "
-                f"Everything is fine"
+            logger.info(
+                f"No capacity for carrier '{carrier}' needed to be"
+                f" distributed. Everything is fine"
             )
 
         elif (
             sum_input.input_capacity_mw.sum() > 0
             and sum_output.output_capacity_mw.sum() == 0
         ):
-            print(
-                f"Error: Capacity for carrier '{carrier}' was not distributed at all!"
+            logger.info(
+                f"Error: Capacity for carrier '{carrier}' was not distributed "
+                f"at all!"
             )
 
         elif (
             sum_output.output_capacity_mw.sum() > 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
-                f"Error: Eventhough no input capacity was provided for carrier '{carrier}' a capacity got distributed!"
+            logger.info(
+                f"Error: Eventhough no input capacity was provided for carrier"
+                f"'{carrier}' a capacity got distributed!"
             )
 
         else:
@@ -169,13 +183,14 @@ def sanitycheck_eGon2035_electricity():
             ) * 100
             g = sum_input["error"].values[0]
 
-            print(f"{carrier}: " + str(round(g, 2)) + " %")
+            logger.info(f"{carrier}: " + str(round(g, 2)) + " %")
 
     # Section to check storage units
 
-    print(f"Sanity checks for scenario {scn}")
-    print(
-        "For German electrical storage units the following deviations between the inputs and outputs can be observed:"
+    logger.info(f"Sanity checks for scenario {scn}")
+    logger.info(
+        "For German electrical storage units the following deviations between"
+        "the inputs and outputs can be observed:"
     )
 
     carriers_electricity = ["pumped_hydro"]
@@ -211,24 +226,27 @@ def sanitycheck_eGon2035_electricity():
             sum_output.output_capacity_mw.sum() == 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
-                f"No capacity for carrier '{carrier}' needed to be distributed. Everything is fine"
+            logger.info(
+                f"No capacity for carrier '{carrier}' needed to be "
+                f"distributed. Everything is fine"
             )
 
         elif (
             sum_input.input_capacity_mw.sum() > 0
             and sum_output.output_capacity_mw.sum() == 0
         ):
-            print(
-                f"Error: Capacity for carrier '{carrier}' was not distributed at all!"
+            logger.info(
+                f"Error: Capacity for carrier '{carrier}' was not distributed"
+                f" at all!"
             )
 
         elif (
             sum_output.output_capacity_mw.sum() > 0
             and sum_input.input_capacity_mw.sum() == 0
         ):
-            print(
-                f"Error: Eventhough no input capacity was provided for carrier '{carrier}' a capacity got distributed!"
+            logger.info(
+                f"Error: Eventhough no input capacity was provided for carrier"
+                f" '{carrier}' a capacity got distributed!"
             )
 
         else:
@@ -238,16 +256,18 @@ def sanitycheck_eGon2035_electricity():
             ) * 100
             g = sum_input["error"].values[0]
 
-            print(f"{carrier}: " + str(round(g, 2)) + " %")
+            logger.info(f"{carrier}: " + str(round(g, 2)) + " %")
 
     # Section to check loads
 
-    print(
-        "For German electricity loads the following deviations between the input and output can be observed:"
+    logger.info(
+        "For German electricity loads the following deviations between the"
+        " input and output can be observed:"
     )
 
     output_demand = db.select_dataframe(
-        """SELECT a.scn_name, a.carrier,  SUM((SELECT SUM(p) FROM UNNEST(b.p_set) p))/1000000::numeric as load_twh
+        """SELECT a.scn_name, a.carrier,  SUM((SELECT SUM(p)
+        FROM UNNEST(b.p_set) p))/1000000::numeric as load_twh
             FROM grid.egon_etrago_load a
             JOIN grid.egon_etrago_load_timeseries b
             ON (a.load_id = b.load_id)
@@ -265,7 +285,8 @@ def sanitycheck_eGon2035_electricity():
     )["load_twh"].values[0]
 
     input_cts_ind = db.select_dataframe(
-        """SELECT scenario, SUM(demand::numeric/1000000) as demand_mw_regio_cts_ind
+        """SELECT scenario,
+         SUM(demand::numeric/1000000) as demand_mw_regio_cts_ind
             FROM demand.egon_demandregio_cts_ind
             WHERE scenario= 'eGon2035'
             AND year IN ('2035')
@@ -289,10 +310,10 @@ def sanitycheck_eGon2035_electricity():
 
     e = round((output_demand - input_demand) / input_demand, 2) * 100
 
-    print(f"electricity demand: {e} %")
+    logger.info(f"electricity demand: {e} %")
 
 
-def sanitycheck_eGon2035_heat():
+def etrago_eGon2035_heat():
     """Execute basic sanity checks.
 
     Returns print statements as sanity checks for the heat sector in
@@ -313,15 +334,18 @@ def sanitycheck_eGon2035_heat():
     scn = "eGon2035"
 
     # Section to check generator capacities
-    print(f"Sanity checks for scenario {scn}")
-    print(
-        "For German heat demands the following deviations between the inputs and outputs can be observed:"
+    logger.info(f"Sanity checks for scenario {scn}")
+    logger.info(
+        "For German heat demands the following deviations between the inputs"
+        " and outputs can be observed:"
     )
 
     # Sanity checks for heat demand
 
     output_heat_demand = db.select_dataframe(
-        """SELECT a.scn_name,  (SUM((SELECT SUM(p) FROM UNNEST(b.p_set) p))/1000000)::numeric as load_twh
+        """SELECT a.scn_name,
+          (SUM(
+          (SELECT SUM(p) FROM UNNEST(b.p_set) p))/1000000)::numeric as load_twh
             FROM grid.egon_etrago_load a
             JOIN grid.egon_etrago_load_timeseries b
             ON (a.load_id = b.load_id)
@@ -351,13 +375,13 @@ def sanitycheck_eGon2035_heat():
         * 100
     )
 
-    print(f"heat demand: {e_demand} %")
+    logger.info(f"heat demand: {e_demand} %")
 
     # Sanity checks for heat supply
 
-    print(
-        "For German heat supplies the following deviations between the inputs and "
-        "outputs can be observed:"
+    logger.info(
+        "For German heat supplies the following deviations between the inputs "
+        "and outputs can be observed:"
     )
 
     # Comparison for central heat pumps
@@ -385,7 +409,7 @@ def sanitycheck_eGon2035_heat():
         round((heat_pump_output - heat_pump_input) / heat_pump_output, 2) * 100
     )
 
-    print(f"'central_heat_pump': {e_heat_pump} % ")
+    logger.info(f"'central_heat_pump': {e_heat_pump} % ")
 
     # Comparison for residential heat pumps
 
@@ -417,11 +441,12 @@ def sanitycheck_eGon2035_heat():
         )
         * 100
     )
-    print(f"'residential heat pumps': {e_residential_heat_pump} %")
+    logger.info(f"'residential heat pumps': {e_residential_heat_pump} %")
 
     # Comparison for resistive heater
     resistive_heater_input = db.select_dataframe(
-        """SELECT carrier, SUM(capacity::numeric) as Urban_central_resistive_heater_MW
+        """SELECT carrier,
+         SUM(capacity::numeric) as Urban_central_resistive_heater_MW
             FROM supply.egon_scenario_capacities
             WHERE carrier= 'urban_central_resistive_heater'
             AND scenario_name IN ('eGon2035')
@@ -449,7 +474,7 @@ def sanitycheck_eGon2035_heat():
         * 100
     )
 
-    print(f"'resistive heater': {e_resistive_heater} %")
+    logger.info(f"'resistive heater': {e_resistive_heater} %")
 
     # Comparison for solar thermal collectors
 
@@ -480,12 +505,13 @@ def sanitycheck_eGon2035_heat():
         )
         * 100
     )
-    print(f"'solar thermal collector': {e_solar_thermal} %")
+    logger.info(f"'solar thermal collector': {e_solar_thermal} %")
 
     # Comparison for geothermal
 
     input_geo_thermal = db.select_dataframe(
-        """SELECT carrier, SUM(capacity::numeric) as Urban_central_geo_thermal_MW
+        """SELECT carrier,
+         SUM(capacity::numeric) as Urban_central_geo_thermal_MW
             FROM supply.egon_scenario_capacities
             WHERE carrier= 'urban_central_geo_thermal'
             AND scenario_name IN ('eGon2035')
@@ -508,7 +534,7 @@ def sanitycheck_eGon2035_heat():
         round((output_geo_thermal - input_geo_thermal) / input_geo_thermal, 2)
         * 100
     )
-    print(f"'geothermal': {e_geo_thermal} %")
+    logger.info(f"'geothermal': {e_geo_thermal} %")
 
 
 def sanitycheck_pv_rooftop_buildings():
@@ -1131,3 +1157,137 @@ def sanitycheck_emobility_mit():
     check_model_data_noflex_eGon2035()
 
     print("=====================================================")
+
+
+def residential_electricity_annual_sum(rtol=1e-5):
+    """Sanity check for dataset electricity_demand_timeseries :
+    Demand_Building_Assignment
+
+    Aggregate the annual demand of all census cells at NUTS3 to compare
+    with initial scaling parameters from DemandRegio.
+    """
+
+    df_nuts3_annual_sum = db.select_dataframe(
+        sql="""
+        SELECT dr.nuts3, dr.scenario, dr.demand_regio_sum, profiles.profile_sum
+        FROM (
+            SELECT scenario, SUM(demand) AS profile_sum, vg250_nuts3
+            FROM demand.egon_demandregio_zensus_electricity AS egon,
+             boundaries.egon_map_zensus_vg250 AS boundaries
+            Where egon.zensus_population_id = boundaries.zensus_population_id
+            AND sector = 'residential'
+            GROUP BY vg250_nuts3, scenario
+            ) AS profiles
+        JOIN (
+            SELECT nuts3, scenario, sum(demand) AS demand_regio_sum
+            FROM demand.egon_demandregio_hh
+            GROUP BY year, scenario, nuts3
+              ) AS dr
+        ON profiles.vg250_nuts3 = dr.nuts3 and profiles.scenario  = dr.scenario
+        """
+    )
+
+    np.testing.assert_allclose(
+        actual=df_nuts3_annual_sum["profile_sum"],
+        desired=df_nuts3_annual_sum["demand_regio_sum"],
+        rtol=rtol,
+        verbose=False,
+    )
+
+    logger.info(
+        "Aggregated annual residential electricity demand"
+        " matches with DemandRegio at NUTS-3."
+    )
+
+
+def residential_electricity_hh_refinement(rtol=1e-5):
+    """Sanity check for dataset electricity_demand_timeseries :
+    Household Demands
+
+    Check sum of aggregated household types after refinement method
+    was applied and compare it to the original census values."""
+
+    df_refinement = db.select_dataframe(
+        sql="""
+        SELECT refined.nuts3, refined.characteristics_code,
+                refined.sum_refined::int, census.sum_census::int
+        FROM(
+            SELECT nuts3, characteristics_code, SUM(hh_10types) as sum_refined
+            FROM society.egon_destatis_zensus_household_per_ha_refined
+            GROUP BY nuts3, characteristics_code)
+            AS refined
+        JOIN(
+            SELECT t.nuts3, t.characteristics_code, sum(orig) as sum_census
+            FROM(
+                SELECT nuts3, cell_id, characteristics_code,
+                        sum(DISTINCT(hh_5types))as orig
+                FROM society.egon_destatis_zensus_household_per_ha_refined
+                GROUP BY cell_id, characteristics_code, nuts3) AS t
+            GROUP BY t.nuts3, t.characteristics_code    ) AS census
+        ON refined.nuts3 = census.nuts3
+        AND refined.characteristics_code = census.characteristics_code
+    """
+    )
+
+    np.testing.assert_allclose(
+        actual=df_refinement["sum_refined"],
+        desired=df_refinement["sum_census"],
+        rtol=rtol,
+        verbose=False,
+    )
+
+    logger.info("All Aggregated household types match at NUTS-3.")
+
+
+def cts_electricity_demand_share(rtol=1e-5):
+    """Sanity check for dataset electricity_demand_timeseries :
+    CtsBuildings
+
+    Check sum of aggregated cts electricity demand share which equals to one
+    for every substation as the substation profile is linearly disaggregated
+    to all buildings."""
+
+    with db.session_scope() as session:
+        cells_query = session.query(EgonCtsElectricityDemandBuildingShare)
+
+    df_demand_share = pd.read_sql(
+        cells_query.statement, cells_query.session.bind, index_col=None
+    )
+
+    np.testing.assert_allclose(
+        actual=df_demand_share.groupby(["bus_id", "scenario"])[
+            "profile_share"
+        ].sum(),
+        desired=1,
+        rtol=rtol,
+        verbose=False,
+    )
+
+    logger.info("The aggregated demand shares equal to one!.")
+
+
+def cts_heat_demand_share(rtol=1e-5):
+    """Sanity check for dataset electricity_demand_timeseries
+    : CtsBuildings
+
+    Check sum of aggregated cts heat demand share which equals to one
+    for every substation as the substation profile is linearly disaggregated
+    to all buildings."""
+
+    with db.session_scope() as session:
+        cells_query = session.query(EgonCtsHeatDemandBuildingShare)
+
+    df_demand_share = pd.read_sql(
+        cells_query.statement, cells_query.session.bind, index_col=None
+    )
+
+    np.testing.assert_allclose(
+        actual=df_demand_share.groupby(["bus_id", "scenario"])[
+            "profile_share"
+        ].sum(),
+        desired=1,
+        rtol=rtol,
+        verbose=False,
+    )
+
+    logger.info("The aggregated demand shares equal to one!.")